add: JwtToken 회원 email, refreshToken 유효기간 추가
